# Preserve proxy invariants