--- a/test.js
+++ b/test.js
@@ -49,37 +49,73 @@ test('works with an array too', t => {
 	const array = onChange(fixture, () => {
 		callCount++;
 	});
-
 	array[0] = 'a';
 	t.deepEqual(array, ['a', 2, {a: false}]);
 	t.is(callCount, 1);
 
 	array[2].a = true;
+	t.deepEqual(array, ['a', 2, {a: true}]);
 	t.is(callCount, 2);
 
 	array.sort();
+	t.deepEqual(array, [2, {a: true}, 'a']);
 	t.is(callCount, 3);
 
 	array.pop();
+	t.deepEqual(array, [2, {a: true}]);
 	t.is(callCount, 4);
 
 	array[2] = false;
+	t.deepEqual(array, [2, {a: true}, false]);
 	t.is(callCount, 5);
 
 	array.reverse();
+	t.deepEqual(array, [false, {a: true}, 2]);
 	t.is(callCount, 6);
 
 	array.reverse();
+	t.deepEqual(array, [2, {a: true}, false]);
 	t.is(callCount, 7);
+
+	array.splice(1, 1, 'a', 'b');
+	t.deepEqual(array, [2, 'a', 'b', false]);
+	t.is(callCount, 8);
 });
 
-// https://github.com/sindresorhus/on-change/issues/14
-test.failing('Array#splice works', t => {
-	const array = onChange([1, 2, 3], () => {});
-
-	t.notThrows(() => {
-		array.splice(0, 1);
+test('invariants', t => {
+	const fixture = {};
+	Object.defineProperty(fixture, 'nonWritable', {
+		configurable: false,
+		writable: false,
+		value: {a: true}
 	});
+	// eslint-disable-next-line accessor-pairs
+	Object.defineProperty(fixture, 'nonReadable', {
+		configurable: false,
+		set: () => {} // No-Op setter
+	});
+	Object.defineProperty(fixture, 'useAccessor', {
+		configurable: false,
+		set(val) {
+			this._useAccessor = val;
+		},
+		get() {
+			return this._useAccessor;
+		}
+	});
+
+	let callCount = 0;
+
+	const proxy = onChange(fixture, () => {
+		callCount++;
+	});
+
+	t.is(proxy.nonWritable, fixture.nonWritable);
+	t.is(proxy.nonReadable, undefined);
+
+	proxy.useAccessor = 10;
+	t.is(proxy.useAccessor, 10);
+	t.is(callCount, 1);
 });
 
 test.cb('the change handler is called after the change is done', t => {