Preserve invariants specified in

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Proxy/handler/get#Invariants

Based on https://github.com/sindresorhus/on-change/pull/15
Based on https://github.com/sindresorhus/on-change/pull/8

Also add full support for accessor properties.