@kpruden 2018-12-19
FYI here are the before/after benchmark results:

Current:
```
[~/dev/js/on-change] ((v0.2.0)) yarn bench
yarn run v1.12.3
$ matcha bench/bench.js
(node:81845) [DEP0006] DeprecationWarning: child_process: options.customFds option is deprecated. Use options.stdio instead.

                      on-change
         121,966 op/s » object read
         104,761 op/s » nested read
         105,137 op/s » array read
         224,256 op/s » object write
         226,778 op/s » array write

                      native
     150,931,689 op/s » object read
     108,503,215 op/s » nested read
      45,009,621 op/s » array read
      15,083,109 op/s » object write
      15,093,054 op/s » array write


  Suites:  2
  Benches: 10
  Elapsed: 79,214.44 ms
```

This PR:
```
[~/dev/js/on-change] (master) yarn bench
yarn run v1.12.3
$ matcha bench/bench.js
(node:81941) [DEP0006] DeprecationWarning: child_process: options.customFds option is deprecated. Use options.stdio instead.

                      on-change
      11,342,177 op/s » object read
       3,251,704 op/s » nested read
       3,391,367 op/s » array read
         496,780 op/s » object write
         629,198 op/s » array write

                      native
     157,343,375 op/s » object read
      92,595,114 op/s » nested read
      43,377,579 op/s » array read
      14,532,006 op/s » object write
      14,300,064 op/s » array write


  Suites:  2
  Benches: 10
  Elapsed: 81,944.30 ms
```

There is some noise in these results as I ran them on my laptop without really taking much care to ensure it wasn't busy with other processes. Nevertheless the performance improvement is clear :)