@sorodrigo 2018-12-22
Hey! Nice stuff, I don't really understand what the proxyTarget stuff is supposed to do 😅 . I know @sindresorhus libs target Node, but is there a way we can avoid symbols? It's a non polyfill-able feature and not adopted by many browsers.