@kpruden 2018-12-22
There is [no direct way](https://stackoverflow.com/questions/51096547/how-to-get-the-target-of-a-javascript-proxy) to retrieve the `target` of a Proxy object (nor is it even possible to determine if an object is in fact a Proxy). The `proxyTarget` attribute is a workaround to determine if the `value` passed to the `set` trap is a proxied value produced by a previous call to `get`. This allows us to "un-proxy" the value before setting it on the underlying target. The benefit of this is that a program can continue to use the raw object it passed to `onChange` to safely bypass change detection, even on descendant properties.

I can switch to a string property name (like `__onChangeProxyTarget` or something). There is a *slight* risk in this case that this could collide with a property of the same name on the underlying `target`. Symbols are safer because two separately-instantiated values never compare as equal (i.e. `Symbol('foo') !== Symbol('foo')`. With a well-chosen name this risk is negligible however.

Looking at the [browser compatibility](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Symbol#Browser_compatibility) for Symbol, it seems about as well-supported as [Proxy](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Proxy#Browser_compatibility) itself. While there is a [polyfill](https://github.com/GoogleChrome/proxy-polyfill) for Proxy, it doesn't appear to support all of the traps used by `on-change` so I don't think using this library in older browsers is in the cards anyway :)