@sindresorhus 2018-12-22
I’m not interested in making the code worse because of browsers.