@sorodrigo 2018-12-22
Ok I think it makes sense if they already have about the same support by browsers and it’s not a desired target of this lib.

@kpruden ok now I get it thanks! I see that you unwrap the proxied value by doing `value = value[proxyTarget]` but where are you wrapping it with that key? Sorry if I’m bothering you too much. I just want to learn 🙂.