@kpruden 2018-12-24
Happy to explain..

In the `set` trap, my goal is to determine if the `value` provided is a Proxy produced by a previous call to the `get` trap. If it is such a Proxy, any property access will be intercepted by our own `get` trap. So, when I access `value[proxyTarget]` our `get` trap will be invoked, which is coded to return `target` when `property === proxyTarget` (line 46). If the `value` passed to `set` is *not* an `on-change`-created Proxy, then `value[proxyTarget]` is `undefined`.