@kpruden 2019-01-07
@sindresorhus do you have any feedback on this PR? I'd love to get it merged... :)