@sindresorhus 2019-01-10
@kpruden This is looking really good. Thanks for elaborating on the implementation details.